+++ src/App.jsx
@@ -4,8 +4,8 @@ import { AppFooter, AppHeader } from './components';
 import { USER_ROUTE } from './pages/user/constants/routes-path/userRoute.constant';
 import './App.css';
 import LoginPage from './pages/login/LoginPage';
-import Register from './pages/login/register/RegisterPage';
-import PwdRestting from './pages/login/pwd-resetting/Password.resetting.component';
+import Register from './pages/register/RegisterPage';
+import PwdRestting from './pages/pwd-resetting/Password.resetting.component';
 import { Chat } from './pages/chat/ChatMainPage';
 import AccommodationList from './pages/accommodation/AccommodationList.page';
 import Main from './pages/main/Main.page';
