--- src/pages/login/login-input/login-sns-image/LoginSnsButton.jsx
@@ -1,12 +1,0 @@
-export default function SnsButton({ Img, alt, onClick, className }) {
-	return (
-		<button>
-			<img
-				src={Img}
-				alt={alt}
-				onClick={onClick}
-				className={className}
-			/>
-		</button>
-	);
-}
