+++ src/pages/login/sns/LoginSnsnButtons.jsx
@@ -1,0 +1,32 @@
+import SnsButton from './button/Login.Sns.Button.component';
+import kakao from '../resource/kakao-sns-auth-btn.png';
+import naver from '../resource/naver-sns-auth-btn.png';
+import google from '../resource/google-sns-auth-btn.png';
+import { KAKAO_AUTH_URL } from '../loginUtil';
+
+export default function SnsButtons() {
+	const sendLogin = () => {
+		window.location.href = KAKAO_AUTH_URL;
+	};
+
+	return (
+		<div className='login-page bottom'>
+			<div className='sns-logo'>
+				<SnsButton
+					onClick={sendLogin}
+					Img={kakao}
+					alt={'카카오 로그인 이미지'}
+				/>
+				<SnsButton
+					className={'naver-button'}
+					Img={naver}
+					alt={'네이버 로그인 이미지'}
+				/>
+				<SnsButton
+					Img={google}
+					alt={'구글 로그인 이미지'}
+				/>
+			</div>
+		</div>
+	);
+}
