+++ src/pages/pwd-resetting/password-resetting/Password-resetting.conponent.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import './password-resetting.conponent.scss';
 import { ResettingInput } from '../resetting-input/resetting.input.conponent';
 import { ResttingTitle } from '../resetting-title/Resetting.title.conponent';
-import { ButtonPrimary } from '../../../../components';
+import { ButtonPrimary } from '../../../components';
 import { useIsResetting } from '../state/resetting.state';
 import { Link } from 'react-router-dom';
