+++ src/pages/pwd-resetting/password-resetting/password-resetting.conponent.scss
