+++ src/pages/pwd-resetting/password.resetting.component.scss
