+++ src/pages/pwd-resetting/resetting-Email-auth/Resetting.email.auth.conponent.jsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
 import './resetting.email.auth.conponent.scss';
-import { ButtonPrimary } from '../../../../components';
+import { ButtonPrimary } from '../../../components';
 import { useIsResetting } from '../state/resetting.state';
 import { ResettingInput } from '../resetting-input/resetting.input.conponent';
 import { ResttingTitle } from '../resetting-title/Resetting.title.conponent';
-import { validateEmail } from '../../util/validate';
+import { validateEmail } from '../../register/util/validate';
 
 export function EmailAuth() {
 	const [email, setEmail] = useState(null);
