+++ src/pages/pwd-resetting/resetting-Email-auth/resetting.email.auth.conponent.scss
