+++ src/pages/pwd-resetting/resetting-input/Resetting.input.conponent.jsx
@@ -1,4 +1,4 @@
-import { InputPrimary } from '../../../../components';
+import { InputPrimary } from '../../../components';
 import { ResttingTitle } from '../resetting-title/Resetting.title.conponent';
 
 export function ResettingInput({
