+++ src/pages/pwd-resetting/resetting-title/Resetting.title.conponent.jsx
@@ -1,4 +1,4 @@
-import { classNames } from '../../../../utils';
+import { classNames } from '../../../utils';
 import './resetting.title.conponent.scss';
 
 export function ResttingTitle({ className, text }) {
