+++ src/pages/pwd-resetting/resetting-title/resetting.title.conponent.scss
