+++ src/pages/register/comple/Comple.component.jsx
@@ -1,6 +1,6 @@
 import './comple.component.scss';
 import { Link } from 'react-router-dom';
-import { ButtonPrimary } from '../../../../components';
+import { ButtonPrimary } from '../../../components';
 import { RegisterInfostore, useRegisterStore } from '../RegisterStore';
 
 export default function RegisterComple() {
