+++ src/pages/register/email/email.component.scss
