+++ src/pages/register/nickName/NickName-sample/NickName.sample.jsx
