+++ src/pages/register/nickName/nickName.component.scss
