+++ src/pages/register/password/password.component.scss
