+++ src/pages/register/registerPage.scss
