+++ src/pages/register/title-text/titleText.component.scss
