+++ src/pages/register/util/validate.js
